requirements.txt
decorator==4.0.6
Django==1.9
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.6.1
django-oauth-toolkit==0.10.0
djangorestframework==3.3.2
ipython==4.0.1
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
readline==6.2.4.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wheel==0.24.0
